+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## UNRELEASED
+
+## `reitit-ring`
+
+* updated deps:
+
+```clj
+[ring/ring-core "1.7.0"] is available but we use "1.6.3"
+```
+
 ## 0.2.2 (2018-09-09)
 
 * better documentation for interceptors