+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 ```
 
 * `:options` requests are served for all routes by default with 200 OK to better support things like [CORS](https://en.wikipedia.org/wiki/Cross-origin_resource_sharing)
-  * the default handler is undocumented (for Swagger)
+  * the default handler is not documented in Swagger
   * new router option `:reitit.ring/default-options-handler` to change this behavior. Setting `nil` disables this.
 
 * updated deps: